+++ src/components/ConfigProvider/ConfigProvider.stories.tsx
@@ -1,4 +1,6 @@
-import React, { useState } from 'react';
+import React, { FC, useState } from 'react';
+import { Stories } from '@storybook/addon-docs';
+import { ComponentStory, ComponentMeta } from '@storybook/react';
 import {
     ButtonSize,
     DefaultButton,
@@ -11,13 +13,50 @@ import { CompactPicker } from 'react-color';
 import { ConfigProvider, OcThemeNames, useConfig } from './';
 import { MatchScore } from '../MatchScore';
 import { Spinner } from '../Spinner';
+import { Stack } from '../Stack';
+import { CheckBoxGroup, CheckboxValueType, RadioGroup } from '../Selectors';
+
+const theme: OcThemeNames[] = [
+    'red',
+    'orange',
+    'yellow',
+    'green',
+    'bluegreen',
+    'blue',
+    'violet',
+    'grey',
+];
 
 export default {
-    title: 'ConfigProvider',
-    component: ConfigProvider,
-};
+    title: 'Config Provider',
+    parameters: {
+        docs: {
+            page: (): JSX.Element => (
+                <main>
+                    <article>
+                        <section>
+                            <h1>Config Provider</h1>
+                            <p>
+                                Config provider is a utility that applies
+                                contextual theming to its child components.
+                                Themes can be applied to your entire app, to
+                                specific subtrees, or to individual components.
+                                By default, components use a blue theme. It also
+                                provides mouse vs. keyboard detection to improve
+                                accessibility.
+                            </p>
+                        </section>
+                        <section>
+                            <Stories includePrimary title="" />
+                        </section>
+                    </article>
+                </main>
+            ),
+        },
+    },
+} as ComponentMeta<typeof ConfigProvider>;
 
-const ThemedComponents = () => {
+const ThemedComponents: FC = () => {
     const [customColor, setCustomColor] = useState<string>('');
     const { themeOptions, setThemeOptions } = useConfig();
     const themes: OcThemeNames[] = [
@@ -36,7 +75,6 @@ const ThemedComponents = () => {
         ariaLabel: `Tab ${i}`,
         ...(i === 4 ? { disabled: true } : {}),
     }));
-
     const iconTabs = [1, 2, 3, 4].map((i) => ({
         value: `tab${i}`,
         icon: IconName.mdiCardsHeart,
@@ -45,8 +83,8 @@ const ThemedComponents = () => {
     }));
 
     return (
-        <>
-            <h1>
+        <Stack direction="vertical" gap="xxl">
+            <h1 style={{ marginBottom: 0 }}>
                 Selected Theme:
                 <span
                     style={{
@@ -58,8 +96,8 @@ const ThemedComponents = () => {
                     {themeOptions.name}
                 </span>
             </h1>
-            <div style={{ display: 'flex', alignItems: 'flex-start' }}>
-                <div style={{ marginRight: '40px' }}>
+            <Stack direction="horizontal" gap="m" style={{ marginTop: 0 }}>
+                <div>
                     <p>Predefined</p>
                     <select
                         value={themeOptions.name}
@@ -77,12 +115,11 @@ const ThemedComponents = () => {
                         ))}
                     </select>
                 </div>
-                <br />
                 <div>
                     <p>Custom</p>
                     <CompactPicker
                         color={customColor}
-                        onChange={(color) => {
+                        onChange={async (color) => {
                             setThemeOptions({
                                 name: 'custom',
                                 customTheme: {
@@ -93,128 +130,146 @@ const ThemedComponents = () => {
                         }}
                     />
                 </div>
-            </div>
-            <br />
-            <br />
-            <div style={{ display: 'flex', gap: '16px' }}>
+            </Stack>
+            <Stack direction="horizontal" gap="m">
                 <PrimaryButton
                     ariaLabel="Primary Button"
-                    onClick={click}
-                    size={ButtonSize.Large}
+                    size={ButtonSize.Small}
                     text="Primary Button"
                 />
                 <PrimaryButton
                     ariaLabel="Primary Button"
-                    onClick={click}
-                    size={ButtonSize.Large}
+                    size={ButtonSize.Small}
                     iconProps={{ path: IconName.mdiCardsHeart }}
                 />
                 <PrimaryButton
                     ariaLabel="Primary Button"
-                    onClick={click}
-                    size={ButtonSize.Large}
+                    size={ButtonSize.Small}
                     iconProps={{ path: IconName.mdiCardsHeart }}
                     text="Primary Button"
                 />
-            </div>
-            <br />
-            <br />
-            <div style={{ display: 'flex', gap: '16px' }}>
+            </Stack>
+            <Stack direction="horizontal" gap="m">
                 <SecondaryButton
                     ariaLabel="Secondary Button"
-                    onClick={click}
-                    size={ButtonSize.Large}
+                    size={ButtonSize.Small}
                     text="Secondary Button"
                 />
                 <SecondaryButton
                     ariaLabel="Secondary Button"
                     iconProps={{ path: IconName.mdiCardsHeart }}
-                    onClick={click}
-                    size={ButtonSize.Large}
+                    size={ButtonSize.Small}
                 />
                 <SecondaryButton
                     ariaLabel="Secondary Button"
-                    onClick={click}
-                    size={ButtonSize.Large}
+                    size={ButtonSize.Small}
                     text="Secondary Button"
                     iconProps={{ path: IconName.mdiCardsHeart }}
                 />
-            </div>
-            <br />
-            <br />
-            <div style={{ display: 'flex', gap: '16px' }}>
+            </Stack>
+            <Stack direction="horizontal" gap="m">
                 <DefaultButton
                     ariaLabel="Default Button"
-                    onClick={click}
-                    size={ButtonSize.Large}
+                    size={ButtonSize.Small}
                     text="Default Button"
                 />
                 <DefaultButton
                     ariaLabel="Default Button"
                     iconProps={{ path: IconName.mdiCardsHeart }}
-                    onClick={click}
-                    size={ButtonSize.Large}
+                    size={ButtonSize.Small}
                 />
                 <DefaultButton
                     ariaLabel="Default Button"
                     iconProps={{ path: IconName.mdiCardsHeart }}
-                    onClick={click}
-                    size={ButtonSize.Large}
+                    size={ButtonSize.Small}
                     text="Default Button"
                 />
-            </div>
-            <br />
-            <br />
-            <Tabs onChange={click} value={'tab1'}>
+            </Stack>
+            <Tabs value={'tab1'}>
                 {tabs.map((tab) => (
                     <Tab key={tab.value} {...tab} />
                 ))}
             </Tabs>
-            <br />
-            <br />
-            <Tabs onChange={click} value={'tab1'}>
+            <Tabs value={'tab1'}>
                 {iconTabs.map((tab) => (
                     <Tab key={tab.value} {...tab} />
                 ))}
             </Tabs>
-            <br />
-            <br />
-            <Tabs onChange={click} value={'tab1'} variant={TabVariant.small}>
+            <Tabs value={'tab1'} variant={TabVariant.small}>
                 {tabs.map((tab) => (
                     <Tab key={tab.value} {...tab} />
                 ))}
             </Tabs>
-            <br />
-            <br />
-            <Tabs onChange={click} value={'tab1'} variant={TabVariant.pill}>
+            <Tabs value={'tab1'} variant={TabVariant.pill}>
                 {tabs.map((tab) => (
                     <Tab key={tab.value} {...tab} />
                 ))}
             </Tabs>
-            <br />
-            <br />
             <MatchScore score={3} />
-            <br />
-            <br />
             <Spinner />
-        </>
+            <CheckBoxGroup
+                {...{
+                    value: ['First'],
+                    defaultChecked: ['First'],
+                    items: [
+                        {
+                            name: 'group',
+                            value: 'First',
+                            label: 'First',
+                            id: 'test-1',
+                        },
+                        {
+                            name: 'group',
+                            value: 'Second',
+                            label: 'Second',
+                            id: 'test-2',
+                        },
+                        {
+                            name: 'group',
+                            value: 'Third',
+                            label: 'Third',
+                            id: 'test-3',
+                        },
+                    ],
+                }}
+            />
+            <RadioGroup
+                {...{
+                    ariaLabel: 'Radio Group',
+                    value: 'Radio1',
+                    items: [1, 2, 3].map((i) => ({
+                        value: `Radio${i}`,
+                        label: `Radio${i}`,
+                        name: 'group',
+                        id: `oea2exk-${i}`,
+                    })),
+                }}
+            />
+        </Stack>
     );
 };
 
-export const Theming = () => {
+const DEFAULT_FOCUS_VISIBLE: boolean = true;
+const DEFAULT_FOCUS_VISIBLE_ELEMENT: HTMLElement = document.documentElement;
+
+const Theming_Story: ComponentStory<typeof ConfigProvider> = (args) => {
     return (
-        <ConfigProvider
-            themeOptions={{
-                name: 'blue',
-            }}
-            focusVisibleOptions={{
-                focusVisible: true,
-                focusVisibleElement: document.getElementById('root'),
-            }}
-        >
+        <ConfigProvider {...args}>
             <ThemedComponents />
         </ConfigProvider>
     );
 };
 
-function click(): void {}
+export const Theming = Theming_Story.bind({});
+
+Theming.args = {
+    focusVisibleOptions: {
+        focusVisible: DEFAULT_FOCUS_VISIBLE,
+        focusVisibleElement: DEFAULT_FOCUS_VISIBLE_ELEMENT,
+    },
+    themeOptions: {
+        name: 'blue',
+    },
+    icomoonIconSet: {},
+    children: <ThemedComponents />,
+};