+++ src/components/List/List.stories.tsx
@@ -1,10 +1,46 @@
 import React from 'react';
+import { Stories } from '@storybook/addon-docs';
+import { ComponentStory, ComponentMeta } from '@storybook/react';
 import { List } from './';
 
 export default {
     title: 'List',
-    component: List,
-};
+    parameters: {
+        docs: {
+            page: (): JSX.Element => (
+                <main>
+                    <article>
+                        <section>
+                            <h1>List</h1>
+                            <p>
+                                A list provides a base component for rendering
+                                sets of items. It's agnostic of layout, and
+                                selection management.
+                            </p>
+                            <h2>Best practices</h2>
+                            <h3>Content</h3>
+                            <ul>
+                                <li>
+                                    Use sentence-style capitalization—only
+                                    capitalize the first word.
+                                </li>
+                            </ul>
+                        </section>
+                        <section>
+                            <Stories includePrimary title="" />
+                        </section>
+                    </article>
+                </main>
+            ),
+        },
+    },
+    argTypes: {
+        layout: {
+            options: ['vertical', 'horizontal'],
+            control: { type: 'inline-radio' },
+        },
+    },
+} as ComponentMeta<typeof List>;
 
 interface User {
     name: string;
@@ -18,53 +54,51 @@ const sampleList: User[] = [1, 2, 3, 4, 5].map((i) => ({
     img: '',
 }));
 
-export const Default = () => (
-    <>
-        <h1>List vertical</h1>
-        <List<User>
-            items={sampleList}
-            renderItem={(item) => (
-                <div>
-                    <p>{item.name}</p>
-                    <div>{item.summary}</div>
-                </div>
-            )}
-            header={
-                <div style={{ paddingLeft: '16px' }}>
-                    <h2>Header</h2>
-                </div>
-            }
-            footer={
-                <div style={{ paddingLeft: '16px' }}>
-                    <h3>Footer</h3>
-                </div>
-            }
-            itemStyle={{ padding: '8px 16px' }}
-        />
-        <br />
-        <br />
-        <br />
-        <h1>List horizontal</h1>
-        <List<User>
-            items={sampleList}
-            renderItem={(item) => (
-                <div>
-                    <p>{item.name}</p>
-                    <div>{item.summary}</div>
-                </div>
-            )}
-            header={
-                <div style={{ paddingLeft: '16px' }}>
-                    <h2>Header</h2>
-                </div>
-            }
-            footer={
-                <div style={{ paddingLeft: '16px' }}>
-                    <h3>Footer</h3>
-                </div>
-            }
-            itemStyle={{ padding: '8px 16px' }}
-            layout={'horizontal'}
-        />
-    </>
+const Vertical_Story: ComponentStory<typeof List> = (args) => (
+    <List {...args} />
 );
+
+export const Vertical = Vertical_Story.bind({});
+
+const Horizontal_Story: ComponentStory<typeof List> = (args) => (
+    <List {...args} />
+);
+
+export const Horizontal = Horizontal_Story.bind({});
+
+const listArgs: Object = {
+    items: sampleList,
+    footer: (
+        <div style={{ paddingLeft: '16px' }}>
+            <h3>Footer</h3>
+        </div>
+    ),
+    layout: 'vertical',
+    renderItem: (item: User) => (
+        <div>
+            <p>{item.name}</p>
+            <div>{item.summary}</div>
+        </div>
+    ),
+    header: (
+        <div style={{ paddingLeft: '16px' }}>
+            <h2>Header</h2>
+        </div>
+    ),
+    classNames: 'my-list-class',
+    style: {},
+    itemClassNames: 'my-list-item-class',
+    itemStyle: { padding: '8px 16px' },
+    listType: 'ul',
+    role: '',
+    itemRole: '',
+};
+
+Vertical.args = {
+    ...listArgs,
+};
+
+Horizontal.args = {
+    ...listArgs,
+    layout: 'horizontal',
+};