+++ src/components/Modal/Modal.tsx
@@ -29,6 +29,7 @@ export const Modal: FC<ModalProps> = React.forwardRef(
 
         const headerClasses: string = mergeClasses([
             styles.header,
+            'header4',
             headerClassNames,
         ]);